Title: 앞의 두 수를 더한 수열 2, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240719/앞의 두 수를 더한 수열 2/a-sequence-by-adding-the-previous-two-numbers-2.cpp
@@ -4,7 +4,7 @@
 
 using namespace std;
 
-int nums[MAX_N+1];
+long long nums[MAX_N+1];
 
 int main() {
     // 여기에 코드를 작성해주세요.
